A C memory/cpu hog finder
This C program inspects the output from the ps aux command and finds the
process that consumes the most resources (either cpu cycles or memory).

The program takes a command-line argument specifying a username, and
will print some information about one process owned by this user to
standard output.

The program accepts an optional argument of -m or -c.  If -c is
provided, your program reports the user's process that has the highest
value for CPU usage. If -m is provided, your program reports the user's
process that has the highest memory usage instead. If neither argument
is provided, it reports the process with the highest CPU usage.

The output of your program is one line containing the PID, usage value
(either memory or CPU - whichever is being compared) with exactly one
digit after the decimal, and command, separated by tabs ('\t').